Use paginator for AWS API calls
This changelist replaces the usage of all paginated API calls with a paginator, introduced in `aws-sdk-go-v2`, that simplifies collecting items returned from a paginated operation.

This changelist is actually generated by a program I wrote to rewrite Go source files to use a paginator, that traverses the Go AST looking for functions that have the return type (`[]awstypes.Type, error`), and determines the AWS package to use to construct a paginator. It then rewrites the function body to use a paginator, while keeping the rest of the code unchanged.

The program is available as a [Gist](https://gist.github.com/cPu1/6c26bad8cb5e2231dc8e3af9bfa2e067).